Create INSTALL.md
The installation instructions provided in README.md aren't working. I was able to build cnn and run the code successfully after spending a lot of time. I've made this file to help others who might be interested in checking out this project. :)

Also, the final binary lstm-parse is built in ner-system directory. So, modify your training and decoding commands in the README file.